# backend/src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Radio Admin Backend API",
    "description": "Documentación interactiva de la API para la emisora web",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8443/"
    },
    {
      "url": "https://localhost:8443/"
    }
  ],
  "paths": {
...
    "/api/nowplaying": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "x-forwarded-host",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "host",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-forwarded-proto",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/search": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "x-forwarded-host",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "host",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-forwarded-proto",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/schedule": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "x-forwarded-host",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "host",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-forwarded-proto",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/station/{stationId}/art/{artId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "stationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "artId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "502": {
            "description": "Bad Gateway"
          }
        }
      }
    },
    "/api/requests/{songId}": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "songId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-forwarded-host",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "host",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-forwarded-proto",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  }
}